requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel file handling

# LLM providers
openai>=1.0.0
google-generativeai>=0.3.0

# qianfan>=0.3.0  # For Baidu Qianfan
# zhipuai>=2.0.0  # For Zhipu AI
# volcengine-python-sdk[ark]>=1.0.0  # For Doubao

# HTTP requests
requests>=2.31.0

# Code quality tools
pylint>=3.0.0
radon>=6.0.0

# Logging
colorlog>=6.7.0

# Environment variables
python-dotenv>=1.0.0
